\name{prep_umaps}
\alias{prep_umaps}
\alias{animate_umaps}
\title{
Run, align, order, and plot UMAPs
}
\description{
Ten UMAPs are generated by \code{prep_umaps}, then rotated to align to each
other and ordered by similarity before being returned by the function. The
function \code{animate_umaps} is then used to create an animation of these UMAP
projections.
}
\usage{
prep_umaps(x, n = 10L, seeds = seq_len(n))

animate_umaps(umap_list, color_by)
}
\arguments{
  \item{x}{
A matrix of input data for UMAP projection. Rows are observations and columns
are variables. For example with single cell data, rows may be cells and columns
may be principal coordinates.
}
  \item{n}{
Number of UMAP projections to generate.
}
  \item{seeds}{
Random seeds for random UMAP generation. Should be a vector \code{n} long.
}
  \item{umap_list}{
List of UMAP projections output by \code{prep_umaps}.
}
  \item{color_by}{
Vector, with as many items as there are rows in \code{x}, of a continuous or
categorical variable to use for coloring points in the plot.
}
}
\value{
Returned by \code{prep_umaps} is a list \code{n} long. Each item in the list is
a matrix with two columns and as many rows as there were rows in \code{x},
indicating UMAP coordinates.

Returned by \code{animate_umaps} is a ggplot object using gganimate.
}
\author{
Lindsay V. Clark
}

\examples{
library(DrumSCircle)
library(ggplot2)

data(pbmc_pca)

use_cells <- which(pbmc_clust %in% c("1", "5", "7"))

u <- make_umaps(pbmc_pca[use_cells, 1:10])

p <- animate_umaps(u, pbmc_clust[use_cells])

print(p + labs(color = "Cluster"))
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
